gdb: Add CT_GDB_CROSS_STATIC_LIBSTDCXX configuration
This commit adds a new configuration `CT_GDB_CROSS_STATIC_LIBSTDCXX`,
which enables static linking of libstdc++ and libgcc for the cross-gdb
running on the host.

This configuration is enabled by default because using shared "standard
libraries" can often cause compatibility problems when distributing
toolchain binaries and it is therefore highly desirable to always
static-link them.

Note that this behaviour aligns with other similar symbols such as
`CT_CC_GCC_STATIC_LIBSTDCXX` and `CT_GDB_NATIVE_STATIC_LIBSTDCXX`.

For macOS, GDB is built using the Clang compiler, which is the default
compiler for the macOS hosts, and `libstdc++` and `libgcc` are not
applicable for obvious reasons.

Since macOS provides a fairly well known and controlled execution
environment, there is little point in static linking the C++ runtime
library.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>