cross-gdb: Add LZMA support configuration
This commit adds a new configuration for enabling the LZMA compression
support through the liblzma, which is required by the GDB's "mini
debuginfo" feature.

Note that this configuration is not enabled by default because it is a
niche feature and requires the liblzma, which is currently not built by
the crosstool-ng for the host (gdb also keeps this feature disabled by
default unless liblzma is available for the host).

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>