cross-gdb: Add Guile scripting support configuration
This commit adds a new configuration for enabling the GNU Guile
scripting support through the libguile.

Note that this configuration is not enabled by default because it is a
niche feature and requires the libguile, which is currently not built
by the crosstool-ng for the host (gdb also keeps this feature disabled
by default unless libguile is available for the host).

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>